Move file type utils to their own file
